# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_hse-revizor_rules-service_internal_pkg_router_dto.CreatePolicyDto:
    properties:
      projectId:
        type: string
      rulesIds:
        items:
          type: string
        type: array
    required:
    - projectId
    - rulesIds
    type: object
  github_com_hse-revizor_rules-service_internal_pkg_router_dto.CreateRuleDto:
    properties:
      params:
        type: string
      typeId:
        type: string
    required:
    - params
    - typeId
    type: object
  github_com_hse-revizor_rules-service_internal_pkg_router_dto.GetPolicyDto:
    properties:
      id:
        type: string
      projectId:
        type: string
      rulesIds:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a Swagger documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Rules Service API
  version: "1.0"
paths:
  /policy:
    post:
      consumes:
      - application/json
      description: Creates a new policy for applying rules to a project
      parameters:
      - description: Policy input
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_hse-revizor_rules-service_internal_pkg_router_dto.CreatePolicyDto'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hse-revizor_rules-service_internal_pkg_router_dto.GetPolicyDto'
      summary: Create policy
      tags:
      - Policy
  /policy/{id}:
    delete:
      description: Deletes policy with provided id
      parameters:
      - description: Policy id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hse-revizor_rules-service_internal_pkg_router_dto.GetPolicyDto'
      summary: Delete policy by id
      tags:
      - Policy
    get:
      description: Returns policy model with provided id
      parameters:
      - description: Policy id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hse-revizor_rules-service_internal_pkg_router_dto.GetPolicyDto'
      summary: Get policy by id
      tags:
      - Policy
  /rule:
    post:
      consumes:
      - application/json
      description: In success case returns created rule model. Type must be equal
        (HasFile, HasStringInFile, HasExpectedValueInField, StrictEquality, HasSubstring,
        HasRegexMatch, NoSubstring, NotLongerThan, NotShorterThan, NotEmpty, DoesLLMSayThatRuleIsSatisfied)
      parameters:
      - description: Rule input
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_hse-revizor_rules-service_internal_pkg_router_dto.CreateRuleDto'
      responses:
        "201":
          description: ' '
      summary: Create rule
      tags:
      - Rule
  /rule/{id}:
    delete:
      description: In success case delete rule model with provided id
      parameters:
      - description: Rule id input
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: ' '
      summary: Delete rule by id
      tags:
      - Rule
    get:
      description: In success case returns rule model with provided id
      parameters:
      - description: Rule id input
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: ' '
      summary: Get rule by id
      tags:
      - Rule
swagger: "2.0"